+++ crates/ruma-events/src/room/join_rules.rs
@@ -2,21 +2,11 @@
 //!
 //! [`m.room.join_rules`]: https://spec.matrix.org/latest/client-server-api/#mroomjoin_rules
 
-use std::{borrow::Cow, collections::BTreeMap};
-
-use ruma_common::{
-    serde::{from_raw_json_value, ignore_invalid_vec_items},
-    space::SpaceRoomJoinRule,
-    OwnedRoomId,
-};
+pub use ruma_common::room::{AllowRule, JoinRule, Restricted};
 use ruma_macros::EventContent;
-use serde::{
-    de::{Deserializer, Error},
-    Deserialize, Serialize,
-};
-use serde_json::{value::RawValue as RawJsonValue, Value as JsonValue};
+use serde::{de, Deserialize, Serialize};
 
-use crate::{EmptyStateKey, PrivOwnedStr};
+use crate::EmptyStateKey;
 
 /// The content of an `m.room.join_rules` event.
 ///
@@ -53,7 +43,7 @@ impl RoomJoinRulesEventContent {
 impl<'de> Deserialize<'de> for RoomJoinRulesEventContent {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
-        D: Deserializer<'de>,
+        D: de::Deserializer<'de>,
     {
         let join_rule = JoinRule::deserialize(deserializer)?;
         Ok(RoomJoinRulesEventContent { join_rule })
@@ -80,191 +70,12 @@ impl SyncRoomJoinRulesEvent {
     }
 }
 
-/// The rule used for users wishing to join this room.
-///
-/// This type can hold an arbitrary string. To check for values that are not available as a
-/// documented variant here, use its string representation, obtained through `.as_str()`.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize)]
-#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-#[serde(tag = "join_rule", rename_all = "snake_case")]
-pub enum JoinRule {
-    /// A user who wishes to join the room must first receive an invite to the room from someone
-    /// already inside of the room.
-    Invite,
-
-    /// Users can join the room if they are invited, or they can request an invite to the room.
-    ///
-    /// They can be allowed (invited) or denied (kicked/banned) access.
-    Knock,
-
-    /// Reserved but not yet implemented by the Matrix specification.
-    Private,
-
-    /// Users can join the room if they are invited, or if they meet any of the conditions
-    /// described in a set of [`AllowRule`]s.
-    Restricted(Restricted),
-
-    /// Users can join the room if they are invited, or if they meet any of the conditions
-    /// described in a set of [`AllowRule`]s, or they can request an invite to the room.
-    KnockRestricted(Restricted),
-
-    /// Anyone can join the room without any prior action.
-    Public,
-
-    #[doc(hidden)]
-    #[serde(skip_serializing)]
-    _Custom(PrivOwnedStr),
-}
-
-impl JoinRule {
-    /// Returns the string name of this `JoinRule`
-    pub fn as_str(&self) -> &str {
-        match self {
-            JoinRule::Invite => "invite",
-            JoinRule::Knock => "knock",
-            JoinRule::Private => "private",
-            JoinRule::Restricted(_) => "restricted",
-            JoinRule::KnockRestricted(_) => "knock_restricted",
-            JoinRule::Public => "public",
-            JoinRule::_Custom(rule) => &rule.0,
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for JoinRule {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let json: Box<RawJsonValue> = Box::deserialize(deserializer)?;
-
-        #[derive(Deserialize)]
-        struct ExtractType<'a> {
-            #[serde(borrow)]
-            join_rule: Option<Cow<'a, str>>,
-        }
-
-        let join_rule = serde_json::from_str::<ExtractType<'_>>(json.get())
-            .map_err(Error::custom)?
-            .join_rule
-            .ok_or_else(|| D::Error::missing_field("join_rule"))?;
-
-        match join_rule.as_ref() {
-            "invite" => Ok(Self::Invite),
-            "knock" => Ok(Self::Knock),
-            "private" => Ok(Self::Private),
-            "restricted" => from_raw_json_value(&json).map(Self::Restricted),
-            "knock_restricted" => from_raw_json_value(&json).map(Self::KnockRestricted),
-            "public" => Ok(Self::Public),
-            _ => Ok(Self::_Custom(PrivOwnedStr(join_rule.into()))),
-        }
-    }
-}
-
-impl From<JoinRule> for SpaceRoomJoinRule {
-    fn from(value: JoinRule) -> Self {
-        value.as_str().into()
-    }
-}
-
-/// Configuration of the `Restricted` join rule.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-pub struct Restricted {
-    /// Allow rules which describe conditions that allow joining a room.
-    #[serde(default, deserialize_with = "ignore_invalid_vec_items")]
-    pub allow: Vec<AllowRule>,
-}
-
-impl Restricted {
-    /// Constructs a new rule set for restricted rooms with the given rules.
-    pub fn new(allow: Vec<AllowRule>) -> Self {
-        Self { allow }
-    }
-}
-
-/// An allow rule which defines a condition that allows joining a room.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize)]
-#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-#[serde(untagged)]
-pub enum AllowRule {
-    /// Joining is allowed if a user is already a member of the room with the id `room_id`.
-    RoomMembership(RoomMembership),
-
-    #[doc(hidden)]
-    _Custom(Box<CustomAllowRule>),
-}
-
-impl AllowRule {
-    /// Constructs an `AllowRule` with membership of the room with the given id as its predicate.
-    pub fn room_membership(room_id: OwnedRoomId) -> Self {
-        Self::RoomMembership(RoomMembership::new(room_id))
-    }
-}
-
-/// Allow rule which grants permission to join based on the membership of another room.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-#[serde(tag = "type", rename = "m.room_membership")]
-pub struct RoomMembership {
-    /// The id of the room which being a member of grants permission to join another room.
-    pub room_id: OwnedRoomId,
-}
-
-impl RoomMembership {
-    /// Constructs a new room membership rule for the given room id.
-    pub fn new(room_id: OwnedRoomId) -> Self {
-        Self { room_id }
-    }
-}
-
-#[doc(hidden)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-pub struct CustomAllowRule {
-    #[serde(rename = "type")]
-    rule_type: String,
-    #[serde(flatten)]
-    extra: BTreeMap<String, JsonValue>,
-}
-
-impl<'de> Deserialize<'de> for AllowRule {
-    fn deserialize<D>(deserializer: D) -> Result<AllowRule, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let json: Box<RawJsonValue> = Box::deserialize(deserializer)?;
-
-        // Extracts the `type` value.
-        #[derive(Deserialize)]
-        struct ExtractType<'a> {
-            #[serde(borrow, rename = "type")]
-            rule_type: Option<Cow<'a, str>>,
-        }
-
-        // Get the value of `type` if present.
-        let rule_type =
-            serde_json::from_str::<ExtractType<'_>>(json.get()).map_err(Error::custom)?.rule_type;
-
-        match rule_type.as_deref() {
-            Some("m.room_membership") => from_raw_json_value(&json).map(Self::RoomMembership),
-            Some(_) => from_raw_json_value(&json).map(Self::_Custom),
-            None => Err(D::Error::missing_field("type")),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::collections::BTreeMap;
-
     use assert_matches2::assert_matches;
     use ruma_common::owned_room_id;
 
-    use super::{
-        AllowRule, CustomAllowRule, JoinRule, OriginalSyncRoomJoinRulesEvent, Restricted,
-        RoomJoinRulesEventContent, SpaceRoomJoinRule,
-    };
+    use super::{AllowRule, JoinRule, OriginalSyncRoomJoinRulesEvent, RoomJoinRulesEventContent};
 
     #[test]
     fn deserialize() {
@@ -289,16 +100,14 @@ mod tests {
             ]
         }"#;
         let event: RoomJoinRulesEventContent = serde_json::from_str(json).unwrap();
-        match event.join_rule {
-            JoinRule::Restricted(restricted) => assert_eq!(
-                restricted.allow,
-                &[
-                    AllowRule::room_membership(owned_room_id!("!mods:example.org")),
-                    AllowRule::room_membership(owned_room_id!("!users:example.org"))
-                ]
-            ),
-            rule => panic!("Deserialized to wrong variant: {rule:?}"),
-        }
+        assert_matches!(event.join_rule, JoinRule::Restricted(restricted));
+        assert_eq!(
+            restricted.allow,
+            &[
+                AllowRule::room_membership(owned_room_id!("!mods:example.org")),
+                AllowRule::room_membership(owned_room_id!("!users:example.org"))
+            ]
+        );
     }
 
     #[test]
@@ -324,82 +133,12 @@ mod tests {
     }
 
     #[test]
-    fn roundtrip_custom_allow_rule() {
-        let json = r#"{"type":"org.msc9000.something","foo":"bar"}"#;
-        let allow_rule: AllowRule = serde_json::from_str(json).unwrap();
-        assert_matches!(&allow_rule, AllowRule::_Custom(_));
-        assert_eq!(serde_json::to_string(&allow_rule).unwrap(), json);
-    }
-
-    #[test]
     fn restricted_room_no_allow_field() {
         let json = r#"{"join_rule":"restricted"}"#;
         let join_rules: RoomJoinRulesEventContent = serde_json::from_str(json).unwrap();
         assert_matches!(
             join_rules,
             RoomJoinRulesEventContent { join_rule: JoinRule::Restricted(_) }
-        );
-    }
-
-    #[test]
-    fn invalid_allow_items() {
-        let json = r#"{
-            "join_rule": "restricted",
-            "allow": [
-                {
-                    "type": "m.room_membership",
-                    "room_id": "!mods:example.org"
-                },
-                {
-                    "type": "m.room_membership",
-                    "room_id": ""
-                },
-                {
-                    "type": "m.room_membership",
-                    "room_id": "not a room id"
-                },
-                {
-                    "type": "org.example.custom",
-                    "org.example.minimum_role": "developer"
-                },
-                {
-                    "not even close": "to being correct",
-                    "any object": "passes this test",
-                    "only non-objects in this array": "cause deserialization to fail"
-                }
-            ]
-        }"#;
-        let event: RoomJoinRulesEventContent = serde_json::from_str(json).unwrap();
-
-        assert_matches!(event.join_rule, JoinRule::Restricted(restricted));
-        assert_eq!(
-            restricted.allow,
-            &[
-                AllowRule::room_membership(owned_room_id!("!mods:example.org")),
-                AllowRule::_Custom(Box::new(CustomAllowRule {
-                    rule_type: "org.example.custom".into(),
-                    extra: BTreeMap::from([(
-                        "org.example.minimum_role".into(),
-                        "developer".into()
-                    )])
-                }))
-            ]
-        );
-    }
-
-    #[test]
-    fn join_rule_to_space_room_join_rule() {
-        assert_eq!(SpaceRoomJoinRule::Invite, JoinRule::Invite.into());
-        assert_eq!(SpaceRoomJoinRule::Knock, JoinRule::Knock.into());
-        assert_eq!(
-            SpaceRoomJoinRule::KnockRestricted,
-            JoinRule::KnockRestricted(Restricted::default()).into()
-        );
-        assert_eq!(SpaceRoomJoinRule::Public, JoinRule::Public.into());
-        assert_eq!(SpaceRoomJoinRule::Private, JoinRule::Private.into());
-        assert_eq!(
-            SpaceRoomJoinRule::Restricted,
-            JoinRule::Restricted(Restricted::default()).into()
         );
     }
 }